using Transformations.map for weather

+++ app/src/main/java/com/tinmegali/myweather/models/Clouds.kt
@@ -6,5 +6,5 @@ import com.google.gson.annotations.SerializedName
 data class Clouds(
         @SerializedName("all")
         @Expose
-        var all: Int? = null
+        var all: Double? = null
 )
