+++ app/src/main/java/com/tinmegali/myweather/models/Rain.kt
@@ -6,5 +6,5 @@ import com.google.gson.annotations.SerializedName
 data class Rain(
         @SerializedName("3h")
         @Expose
-        var _3h: Int? = null
+        var _3h: Double? = null
 )
